chore: move `simsx` to `testutil` (#24481)
