+++ testutil/simsx/msg_factory_test.go
