+++ testutil/simsx/registry.go
