+++ testutil/simsx/registry_test.go
