+++ testutil/simsx/slices.go
