+++ testutil/simsx/slices_test.go
