+++ x/bank/simulation/msg_factory.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"slices"
 
-	"github.com/cosmos/cosmos-sdk/simsx"
+	"github.com/cosmos/cosmos-sdk/testutil/simsx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
 )
