+++ x/mint/simulation/msg_factory.go
@@ -5,7 +5,7 @@ import (
 
 	sdkmath "cosmossdk.io/math"
 
-	"github.com/cosmos/cosmos-sdk/simsx"
+	"github.com/cosmos/cosmos-sdk/testutil/simsx"
 	"github.com/cosmos/cosmos-sdk/x/mint/types"
 )
