+++ x/protocolpool/simulation/msg_factory.go
@@ -3,7 +3,7 @@ package simulation
 import (
 	"context"
 
-	"github.com/cosmos/cosmos-sdk/simsx"
+	"github.com/cosmos/cosmos-sdk/testutil/simsx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/protocolpool/types"
 )
